+++ contracts/metatx/interfaces/IForwarderRegistry.sol
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.25;
 
-/// @title Universal Meta-Transactions Forwarder Registry.
-/// @dev Derived from https://github.com/wighawag/universal-forwarder (MIT licence)
+/// @title Meta-Transactions Forwarder Registry.
 interface IForwarderRegistry {
-    /// @notice Checks whether an account is as an approved meta-transaction forwarder for a sender account.
+    /// @notice Checks whether an account is as an approved meta-transaction forwarder for a sender account to a target contract.
     /// @param sender The sender account.
     /// @param forwarder The forwarder account.
-    /// @return isApproved True if `forwarder` is an approved meta-transaction forwarder for `sender`, false otherwise.
-    function isApprovedForwarder(address sender, address forwarder) external view returns (bool isApproved);
+    /// @param target The target contract.
+    /// @return isApproved True if `forwarder` is an approved meta-transaction forwarder for `sender` to `target`, false otherwise.
+    function isApprovedForwarder(address sender, address forwarder, address target) external view returns (bool isApproved);
 }