+++ test/contracts/metatx/ForwarderRegistry.test.js
@@ -7,6 +7,7 @@ const ForwarderApprovalType = {
   ForwarderApproval: [
     {name: 'sender', type: 'address'},
     {name: 'forwarder', type: 'address'},
+    {name: 'target', type: 'address'},
     {name: 'approved', type: 'bool'},
     {name: 'nonce', type: 'uint256'},
   ],
@@ -55,23 +56,45 @@ describe('Meta Transactions', function () {
       });
     });
 
-    describe('removeForwarderApproval(address,bool)', function () {
-      beforeEach(async function () {
-        this.receipt = await this.contract.removeForwarderApproval(this.forwarder.getAddress());
+    describe('setForwarderApproval(address,address,bool)', function () {
+      context('when setting approval', function () {
+        beforeEach(async function () {
+          this.receipt = await this.contract.setForwarderApproval(this.forwarder.getAddress(), this.receiver.getAddress(), true);
+        });
+
+        it('sets the forwarder approval', async function () {
+          expect(await this.contract.isApprovedForwarder(deployer.address, this.forwarder.getAddress(), this.receiver.getAddress())).to.be.true;
+        });
+
+        it('updates the forwarder approval nonce', async function () {
+          expect(await this.contract.getNonce(deployer.address, this.forwarder.getAddress(), this.receiver.getAddress())).to.equal(1);
+        });
+
+        it('emits a ForwarderApproval event', async function () {
+          await expect(this.receipt)
+            .to.emit(this.contract, 'ForwarderApproval')
+            .withArgs(deployer.address, await this.forwarder.getAddress(), this.receiver.getAddress(), true, 0);
+        });
       });
 
-      it('sets the forwarder approval', async function () {
-        expect(await this.contract.isApprovedForwarder(deployer.address, this.forwarder.getAddress())).to.be.false;
-      });
+      context('when unsetting approval', function () {
+        beforeEach(async function () {
+          this.receipt = await this.contract.setForwarderApproval(this.forwarder.getAddress(), this.receiver.getAddress(), false);
+        });
 
-      it('updates the forwarder approval nonce', async function () {
-        expect(await this.contract.getNonce(deployer.address, this.forwarder.getAddress())).to.equal(1);
-      });
+        it('sets the forwarder approval', async function () {
+          expect(await this.contract.isApprovedForwarder(deployer.address, this.forwarder.getAddress(), this.receiver.getAddress())).to.be.false;
+        });
 
-      it('emits a ForwarderApproval event', async function () {
-        await expect(this.receipt)
-          .to.emit(this.contract, 'ForwarderApproval')
-          .withArgs(deployer.address, await this.forwarder.getAddress(), false, 0);
+        it('updates the forwarder approval nonce', async function () {
+          expect(await this.contract.getNonce(deployer.address, this.forwarder.getAddress(), this.receiver.getAddress())).to.equal(1);
+        });
+
+        it('emits a ForwarderApproval event', async function () {
+          await expect(this.receipt)
+            .to.emit(this.contract, 'ForwarderApproval')
+            .withArgs(deployer.address, await this.forwarder.getAddress(), this.receiver.getAddress(), false, 0);
+        });
       });
     });
 
@@ -90,6 +113,7 @@ describe('Meta Transactions', function () {
         const signature = await deployer.signTypedData(this.domain, ForwarderApprovalType, {
           sender: this.signer,
           forwarder: await this.forwarder.getAddress(),
+          target: await this.receiver.getAddress(),
           approved: false, // should be true
           nonce: 0,
         });
@@ -97,6 +121,7 @@ describe('Meta Transactions', function () {
         const {data: relayerData} = await this.contract.setForwarderApproval.populateTransaction(
           this.signer,
           this.forwarder.getAddress(),
+          this.receiver.getAddress(),
           true,
           signature,
           isEIP712Signature
@@ -112,6 +137,7 @@ describe('Meta Transactions', function () {
           const signature = await deployer.signTypedData(this.domain, ForwarderApprovalType, {
             sender: this.signer,
             forwarder: await this.forwarder.getAddress(),
+            target: await this.receiver.getAddress(),
             approved: true,
             nonce: 0,
           });
@@ -119,6 +145,7 @@ describe('Meta Transactions', function () {
           const {data: relayerData} = await this.contract.setForwarderApproval.populateTransaction(
             this.signer,
             this.forwarder.getAddress(),
+            this.receiver.getAddress(),
             true,
             signature,
             isEIP712Signature
@@ -127,17 +154,17 @@ describe('Meta Transactions', function () {
         });
 
         it('sets the forwarder approval', async function () {
-          expect(await this.contract.isApprovedForwarder(this.signer, this.forwarder.getAddress())).to.be.true;
+          expect(await this.contract.isApprovedForwarder(this.signer, this.forwarder.getAddress(), this.receiver.getAddress())).to.be.true;
         });
 
         it('updates the forwarder approval nonce', async function () {
-          expect(await this.contract.getNonce(this.signer, this.forwarder.getAddress())).to.equal(1);
+          expect(await this.contract.getNonce(this.signer, this.forwarder.getAddress(), this.receiver.getAddress())).to.equal(1);
         });
 
         it('emits a ForwarderApproval event', async function () {
           await expect(this.receipt)
             .to.emit(this.contract, 'ForwarderApproval')
-            .withArgs(this.signer, await this.forwarder.getAddress(), true, 0);
+            .withArgs(this.signer, await this.forwarder.getAddress(), await this.receiver.getAddress(), true, 0);
         });
 
         it('allows direct forwarding to a receiver contract', async function () {
@@ -148,10 +175,10 @@ describe('Meta Transactions', function () {
       });
     }
 
-    describe('setForwarderApproval(address,address,bool,bytes,bool) DIRECT', function () {
+    describe('setForwarderApproval(address,address,address,bool,bytes,bool) DIRECT', function () {
       describeSetForwarderApproval(false);
     });
-    describe('setForwarderApproval(address,address,bool,bytes,bool) ERC1271', function () {
+    describe('setForwarderApproval(address,address,address,bool,bytes,bool) ERC1271', function () {
       describeSetForwarderApproval(true);
     });
 
@@ -160,6 +187,7 @@ describe('Meta Transactions', function () {
         const signature = await deployer.signTypedData(this.domain, ForwarderApprovalType, {
           sender: deployer.address,
           forwarder: await this.forwarder.getAddress(),
+          target: await this.receiver.getAddress(),
           approved: false, // should be true
           nonce: 0,
         });
@@ -177,6 +205,7 @@ describe('Meta Transactions', function () {
           const signature = await deployer.signTypedData(this.domain, ForwarderApprovalType, {
             sender: deployer.address,
             forwarder: await this.forwarder.getAddress(),
+            target: await this.receiver.getAddress(),
             approved: true,
             nonce: 0,
           });
@@ -186,11 +215,11 @@ describe('Meta Transactions', function () {
         });
 
         it('sets the forwarder approval', async function () {
-          expect(await this.contract.isApprovedForwarder(deployer.address, this.forwarder.getAddress())).to.be.true;
+          expect(await this.contract.isApprovedForwarder(deployer.address, this.forwarder.getAddress(), this.receiver.getAddress())).to.be.true;
         });
 
         it('updates the forwarder approval nonce', async function () {
-          expect(await this.contract.getNonce(deployer.address, this.forwarder.getAddress())).to.equal(1);
+          expect(await this.contract.getNonce(deployer.address, this.forwarder.getAddress(), this.receiver.getAddress())).to.equal(1);
         });
 
         it('calls the target function', async function () {
@@ -200,7 +229,7 @@ describe('Meta Transactions', function () {
         it('emits a ForwarderApproval event', async function () {
           await expect(this.receipt)
             .to.emit(this.contract, 'ForwarderApproval')
-            .withArgs(deployer.address, await this.forwarder.getAddress(), true, 0);
+            .withArgs(deployer.address, await this.forwarder.getAddress(), await this.receiver.getAddress(), true, 0);
         });
 
         it('allows direct forwarding to a receiver contract', async function () {
@@ -222,7 +251,7 @@ describe('Meta Transactions', function () {
           })
         )
           .to.be.revertedWithCustomError(this.contract, 'ForwarderNotApproved')
-          .withArgs(deployer.address, other.address);
+          .withArgs(deployer.address, other.address, await this.receiver.getAddress());
       });
 
       context('when successful', function () {
@@ -230,6 +259,7 @@ describe('Meta Transactions', function () {
           const signature = await deployer.signTypedData(this.domain, ForwarderApprovalType, {
             sender: deployer.address,
             forwarder: other.address,
+            target: await this.receiver.getAddress(),
             approved: true,
             nonce: 0,
           });
@@ -237,6 +267,7 @@ describe('Meta Transactions', function () {
           const {data: approvalData} = await this.contract.setForwarderApproval.populateTransaction(
             deployer.address,
             other.address,
+            this.receiver.getAddress(),
             true,
             signature,
             false
@@ -270,6 +301,7 @@ describe('Meta Transactions', function () {
         const signature = await deployer.signTypedData(this.domain, ForwarderApprovalType, {
           sender: deployer.address,
           forwarder: await this.forwarder.getAddress(),
+          target: await this.receiver.getAddress(),
           approved: true,
           nonce: 0,
         });
@@ -277,6 +309,7 @@ describe('Meta Transactions', function () {
         const {data: relayerData} = await this.contract.setForwarderApproval.populateTransaction(
           deployer.address,
           this.forwarder.getAddress(),
+          this.receiver.getAddress(),
           true,
           signature,
           false