+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Breaking changes
 
+- Updated `ForwarderRegistry` so that forwarder approvals are set per meta-tx target contract instead of globally.
+- Replaced `ForwarderRegistry`.`removeForwarderApproval` function with `ForwarderRegistry`.`setForwarderApproval`.
 - Upgraded to `solc@0.8.25`.
 
 ### New features